find revisions to submit
The revisions we want to submit for review are all the revisions between
the current HEAD, and the next branch that we find while walking through
the revisions.  The tip of the branch we find is excluded from the patch
series.

Add some tests as well.  Testing isn't really nice, as there is some
duplicate code for creating a tempdir, but I fought the borrow checker
for too long and couldn't figure it out, so I'll just leave the
duplicate code in for now.