use workdir instead of .git directory
Set the path to the workdir of the repository, instead of setting it to
the .git directory.  This puts the patch files in the working directory
as well, so hopefully nobody has folders named output-$branchname.  It
does however make other operations simpler, so I'll just take it for
now, and maybe change it later.