allow specifying to and cc in the command line
The to and cc addresses need to come from somewhere, and specifying them
on the command line is as good a way as any for now.  As the end-game
the to address might come from a local config, and the cc addresses from
a cc-cmd by default for v1 of a patch series.  For further versions of
patch series they can and should come from gmane.